refactor:separate dukan product event and listener to another file
