+++ dukan/src/main/kotlin/net/thechance/dukan/eventListener/ProductEventListener.kt
@@ -1,0 +1,34 @@
+package net.thechance.dukan.eventListener
+
+import net.thechance.dukan.repository.DukanProductSearchRepository
+import net.thechance.dukan.search.document.ProductDocument
+import net.thechance.events.dukan.ProductEvent
+import org.springframework.context.event.EventListener
+import org.springframework.scheduling.annotation.Async
+import org.springframework.stereotype.Component
+
+@Component
+class ProductEventListener(
+    private val productSearchRepository: DukanProductSearchRepository
+) {
+    @EventListener
+    @Async
+    fun onProductEventListener(event: ProductEvent){
+        when(event){
+            is ProductEvent.Delete -> productSearchRepository.deleteById(event.id)
+            is ProductEvent.Save -> saveProductDocument(event)
+        }
+    }
+
+    private fun saveProductDocument(event: ProductEvent.Save) {
+        val productDocument = ProductDocument(
+            id = event.id,
+            name = event.name,
+            price = event.price,
+            description = event.description,
+            mainImageUrl = event.mainImageUrl,
+            shelfName = event.shelfName
+        )
+        productSearchRepository.save(productDocument)
+    }
+}
