+++ events/src/main/kotlin/net/thechance/events/dukan/DukanEvent.kt
@@ -9,7 +9,7 @@ sealed class DukanEvent : MenaEvent {
         val name: String,
         val status: Status,
         val imageUrl:String?,
-        val location: GeoPoint
+        val location: GeoPoint,
     ):DukanEvent() {
         enum class Status {
             APPROVED,
@@ -20,15 +20,3 @@ sealed class DukanEvent : MenaEvent {
     data class Delete(val id:String):DukanEvent()
 }
 
-sealed class ProductEvent:MenaEvent{
-    data class Save(
-        val id: String,
-        val name: String,
-        val price: Double,
-        val description: String,
-        val mainImageUrl: String,
-        val shelfName: String
-    ):ProductEvent()
-
-    data class Delete(val id:String):ProductEvent()
-}
